Add shared tests for landing page admin and use them for groups and processes
