+++ decidim-participatory_processes/spec/shared/manage_landing_page_examples.rb
@@ -1,0 +1,113 @@
+# frozen_string_literal: true
+
+shared_examples "manage landing page examples" do
+  let(:active_content_blocks) do
+    Decidim::ContentBlock.for_scope(
+      scope_name,
+      organization:
+    ).where(scoped_resource_id: resource.id)
+  end
+
+  before do
+    switch_to_host(organization.host)
+    login_as user, scope: :user
+  end
+
+  context "when editing a participatory process group landing page" do
+    it "has sub nav with Landing page active" do
+      visit edit_landing_page_path
+      within "div.secondary-nav" do
+        expect(page).to have_content("Info")
+        expect(page).to have_content("Landing page")
+        active_secondary_nav = find(:xpath, ".//li[@class='is-active']")
+        expect(active_secondary_nav.text).to eq("Landing page")
+      end
+    end
+  end
+
+  context "when editing a non-persisted content block" do
+    it "creates the content block to the db before editing it" do
+      visit edit_landing_page_path
+
+      expect do
+        within ".js-list-availables" do
+          within find("li", text: "Hero image") do
+            find("svg.icon--pencil").click
+          end
+        end
+      end.to change(active_content_blocks, :count).by(1)
+    end
+
+    it "creates the content block when dragged from inactive to active panel" do
+      visit edit_landing_page_path
+      content_block = first("ul.js-list-availables li")
+      active_blocks_list = find("ul.js-list-actives")
+
+      expect do
+        content_block.drag_to(active_blocks_list)
+        sleep(2)
+      end.to change(active_content_blocks, :count).by(1)
+    end
+  end
+
+  context "when editing a persisted content block" do
+    let!(:content_block) do
+      create(
+        :content_block,
+        organization:,
+        manifest_name: :hero,
+        scope_name:,
+        scoped_resource_id: resource.id
+      )
+    end
+
+    before do
+      switch_to_host(organization.host)
+      login_as user, scope: :user
+    end
+
+    it "updates the settings of the content block" do
+      visit edit_content_block_path(resource, :hero)
+
+      fill_in(
+        :content_block_settings_welcome_text_en,
+        with: "Custom welcome text!"
+      )
+
+      click_button "Update"
+      visit edit_content_block_path(resource, :hero)
+      expect(page).to have_selector("input[value='Custom welcome text!']")
+
+      content_block.reload
+
+      expect(content_block.settings.to_json).to match(/Custom welcome text!/)
+    end
+  end
+end
+
+shared_context "when admin administrating a participatory process with hero content block registered" do
+  include_context "when admin administrating a participatory process"
+
+  before do
+    unless Decidim.content_blocks.for(scope_name).any? { |cb| cb.name == :hero }
+      Decidim.content_blocks.register(scope_name, :hero) do |content_block|
+        content_block.cell = "decidim/content_blocks/hero"
+        content_block.settings_form_cell = "decidim/content_blocks/hero_settings_form"
+        content_block.public_name_key = "decidim.content_blocks.hero.name"
+
+        content_block.images = [
+          {
+            name: :background_image,
+            uploader: "Decidim::HomepageImageUploader"
+          }
+        ]
+
+        content_block.settings do |settings|
+          settings.attribute :welcome_text, type: :text, translated: true
+        end
+
+        content_block.default!
+      end
+    end
+  end
+end
